again

+++ 0x00-hello_world/6-size.c
@@ -13,7 +13,7 @@ char charType;
 printf("size of char: %zu byte(s)\n", sizeof(char));
 printf("size of int: %zu byte(s)\n", sizeof(int));
 printf("size of long int: %zu byte(s)\n", sizeof(long int));
-printf("size of long long int: %zu byte(s)\n", sizeof(longlong int));
+printf("size of long long int: %zu byte(s)\n", sizeof(long long int));
 printf("size of float: %zu byte(s)\n", sizeof(float));
 return (0);
 }
